requirements.txt
asgiref==3.5.0
boto3==1.20.46
botocore==1.23.46
Django==4.0.1
django-environ==0.8.1
django-storages==1.12.3
gunicorn==20.1.0
jmespath==0.10.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
